+++ include/s_user.h
@@ -99,6 +99,7 @@ extern char *umode_str(struct Client *cptr);
 extern void send_umode(struct Client *cptr, struct Client *sptr,
                        struct Flags *old, int sendset);
 extern void set_snomask(struct Client *, unsigned int, int);
+extern const char *snomask_to_str(unsigned int mask, char *buf, size_t buflen);
 extern int is_snomask(char *);
 extern int check_target_limit(struct Client *sptr, void *target, const char *name,
     int created);